requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
httpx==0.28.1
PyJWT==2.10.1
python-docx==1.1.2
python-multipart==0.0.20
google-generativeai==0.8.3
Pillow==11.1.0
